Move metric compatibility tests into focused file
